Work hirota

+++ ProjectTeamB/WEB-INF/src/admin_servlet/CmpDelServlet.java
@@ -21,35 +21,37 @@ public class CmpDelServlet extends HttpServlet {
 	}
 
 	//Compare names→Get_name&Get_pass
-	protected void doGet(HttpServletRequest request, HttpServletResponse response)
+	public void doGet(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 
 		//Initialization
 		request.setCharacterEncoding("UTF-8");
 		response.setContentType("text/html; charset=utf-8");
 		HttpSession session = request.getSession();
-		String[] delname = (String[])session.getAttribute("delname");
-		String botton = "社員一覧ページ";
-		String URL = "/ManagerServlet";
-		
-		StaffDAO stdao = new StaffDAO();
-		int del = stdao.deleteStaff(delname);
-		
-		if(del== delname.length) {
-			request.setAttribute("botton",botton);
-			request.setAttribute("URL",URL);
-			RequestDispatcher rd = request.getRequestDispatcher("/view/VersView/Complete.jsp");
+		String name = (String) session.getAttribute("name");
+		if (name == null) {
+			RequestDispatcher rd = request.getRequestDispatcher("/view/VersView/Timeout.jsp");
 			rd.forward(request, response);
-		}else {
-			request.setAttribute("botton",botton);
-			request.setAttribute("URL",URL);
-			RequestDispatcher rd = request.getRequestDispatcher("/view/VersView/Error.jsp");
-			rd.forward(request, response);
+		} else {
+			String[] delname = (String[]) session.getAttribute("delname");
+			String botton = "社員一覧ページ";
+			String URL = "/ManagerServlet";
+
+			StaffDAO stdao = new StaffDAO();
+			int del = stdao.deleteStaff(delname);
+
+			if (del == delname.length) {
+				request.setAttribute("botton", botton);
+				request.setAttribute("URL", URL);
+				RequestDispatcher rd = request.getRequestDispatcher("/view/VersView/Complete.jsp");
+				rd.forward(request, response);
+			} else {
+				request.setAttribute("botton", botton);
+				request.setAttribute("URL", URL);
+				RequestDispatcher rd = request.getRequestDispatcher("/view/VersView/Error.jsp");
+				rd.forward(request, response);
+			}
 		}
 	}
 
-	protected void doPost(HttpServletRequest request, HttpServletResponse response)
-			throws ServletException, IOException {
-		doGet(request, response);
-	}
 }
